[kernel] E: Refactor mm into feature-gated modules
Extract the memory manager initialization logic into
separate modules gated by platform-root-virtual-address-space-bootstrap:

- kernel_vas.rs: kernel-built root VAS path (identity-map, CR3 load,
  MMIO mapping). Moved from mm/mod.rs.
- virt/boot_init.rs: boot-time page table construction. Moved from
  virt/mod.rs.

This separation allows platforms that inherit a pre-built VAS from the
host (e.g., Hyperlight i686-guest) to compile without the identity-map
code path, while keeping the microvm path unchanged.

Other changes:
- Re-export virt::memcpy under cfg(feature = "hyperlight") so that
  crate::mm::memcpy is accessible from the hyperlight platform module.
- Make phys module pub(crate) for cross-module frame allocation.
- Extend page_directory and page_table with helper methods used by the
  new boot_init path.
- Add VAS to codespell ignore list (Virtual Address Space acronym).
