+++ src/kernel/src/hal/mem/types/address/mod.rs
@@ -25,6 +25,7 @@ pub use ::sys::mm::{
 pub use aligned::*;
 pub use frame::*;
 pub use page::*;
+#[cfg(not(feature = "platform-root-virtual-address-space-bootstrap"))]
 pub use pd::*;
 pub use phys::*;
